# docs/swagger.yaml
basePath: /v1
definitions:
...
  reportdomain.ChartData:
    properties:
      hour:
        type: integer
      label:
        type: string
      value:
        type: number
    type: object
...
  reportdomain.CohortRetention:
    properties:
      cohort_month:
        description: e.g., "Jan 2024"
        type: string
      cohort_size:
        description: Total users in cohort
        type: integer
      retention:
        description: Percentage for Month 0, 1, 2...
        items:
          type: number
        type: array
    type: object
...
  reportdomain.StackedChartData:
    properties:
      label:
        description: Month Name
        type: string
      new:
        description: Registered this month
        type: integer
      recurring:
        description: Active this month but registered previously
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for VitalFit, a gym system management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VitalFit API
paths:
...
  /reports/charts/cohort-analysis:
    get:
      description: Retrieves user retention percentages grouped by registration month
        (cohort). Month 0 is always 100%. Subsequent months show the % of users who
        made a payment.
      parameters:
      - description: Filter by Branch UUID
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cohort analysis data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.CohortRetention'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Cohort Analysis (Retention Heatmap)
      tags:
      - Reports
...
  /reports/charts/most-used-services:
    get:
      description: Retrieves the top 5 most used services based on attendance.
      parameters:
      - description: Start date for the report (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date for the report (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Most used services data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.ChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Most Used Services
      tags:
      - Reports
  /reports/charts/new-vs-recurring:
    get:
      description: Retrieves the evolution of New vs Recurring users per month for
        the current year. Ideal for Stacked Area Charts.
      parameters:
      - description: Filter by Branch UUID
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: New vs Recurring data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.StackedChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get New vs Recurring Users Chart
      tags:
      - Reports
  /reports/charts/sales-by-category:
    get:
      description: Retrieves sales data grouped by category (Memberships, Packages,
        Service Categories) ordered by revenue. Ideal for Horizontal Bar Charts.
      parameters:
      - description: Filter by Branch UUID
        in: query
        name: branch_id
        type: string
      - description: Start date for the report (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date for the report (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales volume by category data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.ChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Sales Volume By Service (Category)
      tags:
      - Reports
  /reports/charts/sales-by-demographics:
    get:
      description: Retrieves sales volume grouped by demographic dimension (age or
        gender).
      parameters:
      - description: Filter by Branch UUID
        in: query
        name: branch_id
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        type: string
      - description: 'Dimension: ''age'' or ''gender'''
        in: query
        name: dimension
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Demographic sales data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.ChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Sales By Demographics
      tags:
      - Reports
  /reports/charts/sales-by-hour:
    get:
      description: Retrieves sales data grouped by the hour of the day for a Line
        Chart.
      produces:
      - application/json
      responses:
        "200":
          description: Sales by hour data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.ChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Sales By Hour
      tags:
      - Reports
  /reports/charts/sales-by-payment-method:
    get:
      description: Retrieves sales data grouped by payment method for a Pie Chart.
      parameters:
      - description: Start date for the report (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date for the report (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales by payment method data
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.ChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Sales By Payment Method
      tags:
      - Reports
  /reports/charts/top-instructors:
    get:
      description: Retrieves the top 5 instructors with the most class attendances
        for a Bar Chart.
      parameters:
      - description: Start date for the report (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date for the report (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Top instructors by attendance
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/reportdomain.ChartData'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Top Instructors by Attendance
      tags:
      - Reports
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"